Move Composition CRD into a separate package to cut dependency
