+++ pkg/composition/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "client.go",
         "controller.go",
         "controller_constructor.go",
         "finalizers.go",
@@ -15,8 +14,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//:go_default_library",
-        "//pkg/api/schema:go_default_library",
-        "//pkg/apis/composition:go_default_library",
         "//pkg/apis/composition/v1:go_default_library",
         "//pkg/apis/formation/v1:go_default_library",
         "//pkg/composition/client:go_default_library",
@@ -48,7 +45,6 @@ go_library(
         "//vendor/github.com/schollz/closestmatch:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
@@ -63,7 +59,6 @@ go_test(
     name = "go_default_test",
     size = "small",
     srcs = [
-        "client_test.go",
         "controller_test.go",
         "plumbing_for_test.go",
         "sd_var_model_test.go",
@@ -95,10 +90,6 @@ go_test(
         "//vendor/github.com/atlassian/smith/pkg/specchecker:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/store:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
-        "//vendor/github.com/go-openapi/errors:go_default_library",
-        "//vendor/github.com/go-openapi/spec:go_default_library",
-        "//vendor/github.com/go-openapi/strfmt:go_default_library",
-        "//vendor/github.com/go-openapi/validate:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/mock:go_default_library",
