+++ pkg/composition/crd/testdata/test_crd_allows_no_resource_group_resources.yml
