+++ pkg/composition/crd/testdata/test_crd_allows_scope_with_account_no_label.yml
