+++ pkg/composition/crd/testdata/test_crd_requires_resource_group_locations.yml
